scripts/update_charts/requirements.txt
requests>=2.26.0
pathlib>=1.0.1
typing>=3.7.4.3
argparse>=1.4.0
PyYAML>=6.0.1
textwrap3>=0.9.2
semver>=3.0.2
